Refactored ed25519 into ed25519 and edwards25519
