+++ Cargo.toml
@@ -33,6 +33,7 @@ members = [
     "examples/gf128",
     "examples/aes128-gcm",
     "examples/bls12-381-hash",
+    "examples/edwards25519",
     "examples/ed25519",
     "examples/no_std",
 ]
@@ -66,6 +67,7 @@ default-members = [
     "examples/gf128",
     "examples/aes128-gcm",
     "examples/bls12-381-hash",
+    "examples/edwards25519",
     "examples/ed25519",
     "examples/sha512",
 ]
