+++ examples/ed25519/tests/test_ed25519.rs
@@ -1,5 +1,6 @@
 use hacspec_dev::prelude::*;
 use hacspec_ed25519::*;
+use hacspec_edwards25519::*;
 use hacspec_lib::*;
 use quickcheck::QuickCheck;
