+++ examples/edwards25519/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hacspec-ed25519"
+name = "hacspec-edwards25519"
 version = "0.1.0"
 authors = ["Kasper S. Nielsen <ksn@cs.au.dk>"]
 edition = "2018"
@@ -7,18 +7,8 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [lib]
-path = "src/ed25519.rs"
+path = "src/edwards25519.rs"
 
 [dependencies]
 hacspec-lib = { path = "../../lib" }
 hacspec-sha512 = { path = "../sha512" }
-hacspec-edwards25519 = { path = "../edwards25519" }
-
-[dev-dependencies]
-quickcheck = "1"
-quickcheck_macros = "1"
-serde_json = "1.0"
-serde = {version = "1.0", features = ["derive"]}
-hacspec-dev = { path = "../../utils/dev" }
-
-
